Send LSP-compliant responses to codelens requests
In the responses to textDocument/codeLens and codeLens/resolve,
go-language-server currently returns objects with unnecessary fields
that are not specified in the Language Server Protocol.  This is not a
big problem per-se, because the extra JSON fields can be ignored by the
consumers.  However, they represent a significant amount of data.

In particular, each item in the CodeLens array that is returned as the
response to textDocument/codeLens contains the text of the whole
document.  As an example, I opened the file ./src/net/http/h2_bundle.go
from the Go source code [1].  Its size is 310K and contains about 440
functions, meaning we'll return about 440 code lenses.  Just the
duplicated in each entry that contains the document text will contribute
for 136M of the response (the total response is bigger).

This is because the code returns blindly the internal vscode data
structures.  Because the CodeLens vscode data structure contains a
reference to the text document, it is included every time the structure
is serialized.  Instead, we should generate structures that only
contain the fields specified in the LSP.

[1] https://github.com/golang/go/blob/master/src/net/http/h2_bundle.go