Call server.initialized() in test before function
If initialized is not called, all the requests to the server block on
the

    await this.activation;

line that is at the beginning of the handlers.  Without this change, all
the tests end up timing out because of that.

Also, the initialized notification in the LSP does not have any
parameters, so remove it from the server method as well.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>