Don't override console.{log,info,warn,error} in tests
Currently, the test output (when running npm run test) does not appear
anywhere.  This is because we pass a ConsoleLogger with logEnabled =
false to LspServer.  This loggers gets wrapped in a PrefixingLogger, and
console.log is replaced with PrefixingLogger.log.  The result is that
when console.log is called, it calls PrefixingLogger.log, then
ConsoleLogger.log, which does nothing.

In the context of the unit tests, I don't think it makes sense to mess
with the output (either prefix or redirect it), so console.log (and
error, warn, info) should not be modified.

This patch makes the caller of LspServer() responsible of instanciating
the PrefixingLogger (if it wants to).  The logger is passed directly to
the activate function, which will override the
console.{log,info,warn,error} functions.  The tests pass "console" as
the logger, which effectively leaves the console object unchanged.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>